Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/scripthash-basic-blocked-error-event-expected.txt

CONSOLE MESSAGE: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'self' 'sha256-deadbeef'". Either the 'unsafe-inline' keyword, a hash ('sha256-ZbQewvwidB/diiqNqVYuj5TxIArbdd1TUb2cl6rIyyg='), or a nonce ('nonce-...') is required to enable inline execution.

This is a testharness.js-based test.
PASS CSP script-hash block causes error event
Harness: the test ran to completion.
